Reorganize mesh coordinate creation into layout subpackage
Move coordinate-creation functions from coords.py and connectivity/triangular.py into layout/rectilinear.py and layout/triangular.py. Drop triangular-specific convenience wrappers in favour of the generic two-step API. Retain backward-compatible re-exports in coords.py and connectivity/triangular.py.

Update all imports and rewrite tests to use the two-step API.
