+++ src/weather_model_graphs/create/mesh/__init__.py
@@ -1,15 +1,15 @@
+from .connectivity.general import create_directed_mesh_graph
 from .connectivity.triangular import (
     create_flat_multiscale_from_triangular_coordinates,
-    create_flat_multiscale_triangular_mesh_graph,
-    create_flat_singlescale_triangular_mesh_graph,
-    create_hierarchical_triangular_mesh_graph,
-    create_multirange_2d_triangular_mesh_primitives,
-    create_single_level_2d_triangular_mesh_graph,
-    create_single_level_2d_triangular_mesh_primitive,
 )
-from .coords import (
-    create_directed_mesh_graph,
+from .layout.rectilinear import (
+    create_multirange_2d_mesh_graphs,
     create_multirange_2d_mesh_primitives,
     create_single_level_2d_mesh_graph,
     create_single_level_2d_mesh_primitive,
+)
+from .layout.triangular import (
+    create_multirange_2d_triangular_mesh_primitives,
+    create_single_level_2d_triangular_mesh_graph,
+    create_single_level_2d_triangular_mesh_primitive,
 )
